+++ ext/posix/posix.c
@@ -10,6 +10,7 @@
  * With documentation from Steve Donovan 2012
  */
 
+#include "bit32.c"
 #include "ctype.c"
 #include "dirent.c"
 #include "errno.c"